Update Card System
Change Show Ability Descriptions & if ability cards are enough, show different cards

+++ Riot_Proto/Assets/Script/Ability/IncreaseAS.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class IncreaseAS : AbilityBase
 {
 
-    public int increaseValue;
+    public float increaseValue;
 
     bool isOne = false;
